Fix phases 2-5 audit bugs with tests
Confirmed and fixed:
- BUG-15: __CxxPureMSILEntry passes envp (Ldarg_2) for 3-param main
- BUG-16: Pointer cast uses conv.i/conv.u (native int), not conv.i8;
  bool-to-pointer path also uses conv.i
- BUG-18: Parser case constant truncation (int)→long; codegen emits
  EmitConstI8 for 64-bit switch conditions
- BUG-23: GenAddr FunCall spills struct-returning calls to scratch
  local so f().member works
- Scratch local type-safety: GetOrAddScratchLocal now checks type
  identity (via GetTypeId) for struct/union/array, preventing IL
  verifier rejection from reusing slots across distinct value types

7 xUnit tests in AuditPhase2Tests.cs.

+++ chibil/Parser.cs
@@ -1151,9 +1151,9 @@ public class Parser
         {
             if (_currentSwitch == null) Util.ErrorTok(tok, "stray case");
             var node = NewNode(NodeKind.Case, tok);
-            int begin = (int)ConstExpr(ref tok, tok.Next);
-            int end;
-            if (Util.Equal(tok, "...")) { end = (int)ConstExpr(ref tok, tok.Next); if (end < begin) Util.ErrorTok(tok, "empty case range specified"); }
+            long begin = ConstExpr(ref tok, tok.Next);
+            long end;
+            if (Util.Equal(tok, "...")) { end = ConstExpr(ref tok, tok.Next); if (end < begin) Util.ErrorTok(tok, "empty case range specified"); }
             else end = begin;
             tok = Util.Skip(tok, ":"); node.Label = NewUniqueName(); node.Lhs = Stmt(ref rest, tok);
             node.Begin = begin; node.End = end; node.CaseNext = _currentSwitch.CaseNext; _currentSwitch.CaseNext = node; return node;
